Move errors into their own module
For #102
