+++ git-mit-config/src/main.rs
@@ -1,20 +1,14 @@
-use std::{
-    env,
-    error::Error,
-    fmt::{Display, Formatter},
-    io,
-};
+use std::{env, io};
 
 use clap::{crate_authors, crate_version, App, AppSettings, Arg};
 use git2::{Config, Repository};
 
+use crate::errors::GitMitConfigError;
 use crate::lints::manage_lints;
 use clap_generate::generate;
 use clap_generate::generators::{Bash, Elvish, Fish, Zsh};
 use mit_commit_message_lints::lints::Lint;
-use mit_commit_message_lints::{
-    author::entities::Authors, errors::MitCommitMessageLintsError, external::vcs::Git2,
-};
+use mit_commit_message_lints::{author::entities::Authors, external::vcs::Git2};
 use std::convert::TryInto;
 
 const LOCAL_SCOPE: &str = "local";
@@ -144,53 +138,4 @@ fn app() -> App<'static> {
 
 mod lints;
 
-#[derive(Debug)]
-pub enum GitMitConfigError {
-    ConfigIoGit2Error(String),
-    UnrecognisedLintCommand,
-    LintNameNotGiven,
-    MitCommitMessageLintsError(MitCommitMessageLintsError),
-    Io(String, String),
-}
-
-impl Display for GitMitConfigError {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            GitMitConfigError::UnrecognisedLintCommand => write!(
-                f,
-                "Unrecognised Lint command, only you may only enable or disable, or list \
-                 available lints"
-            ),
-            GitMitConfigError::LintNameNotGiven => write!(f, "Please specify a lint"),
-            GitMitConfigError::MitCommitMessageLintsError(error) => write!(f, "{}", error),
-            GitMitConfigError::Io(file_source, error) => write!(
-                f,
-                "Failed to read git config from `{}`:\n{}",
-                file_source, error
-            ),
-            GitMitConfigError::ConfigIoGit2Error(error) => {
-                write!(f, "Failed to load a git config: {}", error)
-            }
-        }
-    }
-}
-
-impl From<git2::Error> for GitMitConfigError {
-    fn from(error: git2::Error) -> Self {
-        GitMitConfigError::ConfigIoGit2Error(format!("{}", error))
-    }
-}
-
-impl From<MitCommitMessageLintsError> for GitMitConfigError {
-    fn from(from: MitCommitMessageLintsError) -> Self {
-        GitMitConfigError::MitCommitMessageLintsError(from)
-    }
-}
-
-impl Error for GitMitConfigError {}
-
-impl GitMitConfigError {
-    fn new_io(source: String, error: &std::io::Error) -> GitMitConfigError {
-        GitMitConfigError::Io(source, format!("{}", error))
-    }
-}
+mod errors;
